Expose Headers, Response and Request constructor as exports
I'm now in two minds whether this should be done at all. In the browser I think no - isomorphic fetch should be fairly pure a fetch polyfill, in my [not particularly strong] opinion

In which case is it really a good idea to have a different API on the server?