Fix "self is not defined" on Browserified bundle running Node
If you browserify a bundle that includes isomorphic-fetch, it only adds the client-side fetch (whatwg-fetch) to the bundle. 
As a result, when importing the browserified bundle via Node, it fails to import "fetch-npm-node" and throws the error "self is not defined".

Self cannot be assumed to be defined.

In reality, a check should also be made to see if `window` exists as a backup, but I made the PR as minimal as possible.